Test that encoding results of network messages do not change
